requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.15.0
plotly>=5.18.0
dash>=2.14.0
geopy>=2.4.0
pyproj>=3.6.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0
